Refactor policy resources
