+++ tests/suite/test_rl_policies.py
@@ -6,6 +6,8 @@ from suite.custom_resources_utils import (
     delete_virtual_server,
     create_virtual_server_from_yaml,
     patch_virtual_server_from_yaml,
+)
+from suite.policy_resources_utils import (
     create_policy_from_yaml,
     delete_policy,
     read_policy,
