+++ tests/suite/test_rl_policies_vsr.py
@@ -7,6 +7,8 @@ from suite.custom_resources_utils import (
     create_virtual_server_from_yaml,
     patch_virtual_server_from_yaml,
     patch_v_s_route_from_yaml,
+)
+from suite.policy_resources_utils import (
     create_policy_from_yaml,
     delete_policy,
     read_policy,
